pyproject.toml
[tool.poetry]
name = "qa-demo-repository"
version = "1.3.0"
description = "Automated UI/API demo testing setup"
authors = ["Jacek Gwoździewicz"]

packages = [
    { include = "qa_demo_repository", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pytest = "^8.3.5"
pytest-playwright = "^0.5.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pytest-xdist = "^3.6.0"
pytest-html = "^4.1.1"
pre-commit = "^4.2.0"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
faker = "^37.3.0"
commitizen = "^4.8.3"


[tool.poetry.scripts]
lints = "qa_demo_repository.cli:lints"

[tool.pytest.ini_options]
markers = [
    "ui: UI ",
    "api: API ",
    "usertests:  usertests ",
    "product_details:  product_details ",
    "cart:  cart ",
    "shopping_modal:  shopping_modal "
]
addopts = "--color=yes --capture=tee-sys"
filterwarnings = "ignore:Unverified HTTPS request.*"


[tool.black]
line-length = 88
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = "qa_demo_repository"

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.3.0"
tag_format = "$version"
changelog_file = "Changelog.md"
update_changelog_on_bump = true
changelog_incremental = true
version_files = ["pyproject.toml:version", "Changelog.md"]
